finished merge sort
The method of visualizing the merge sort is different from the others. It's probably better and I might change the other algorithms to use the same system. Anyway, the new system sorts the array, logs every change, and then plays it back. Whereas, with the old method, the array would be sorted live and each change would have a delay attached to it.